finalizando feature de inscrever participantes e liberar a alimentação por planilha
